Move path resolution logic to `Loader`.
