+++ packages/core-kit/src/nodes/invoke.ts
@@ -23,26 +23,11 @@ export type InvokeNodeInputs = InputValues & {
   graph?: GraphDescriptor;
 };
 
-// This looks like a generic utility function that could be moved to a shared
-// location.
-export const baseURLFromContext = (context: NodeHandlerContext) => {
-  if (context.outerGraph?.url) return new URL(context.outerGraph.url);
-  const invokingBoardURL = context.board?.url;
-  if (invokingBoardURL) return new URL(invokingBoardURL);
-  if (context.base) return context.base;
-  // This should probably return SENTINEL_BASE_URL.
-  return new URL(import.meta.url);
-};
-
 export const loadBoardFromPath = async (
   path: string,
   context: NodeHandlerContext
 ) => {
-  const base = baseURLFromContext(context);
-  const graph = await context.loader?.load(path, {
-    base,
-    outerGraph: context.outerGraph,
-  });
+  const graph = await context.loader?.load(path, context);
   if (!graph) throw new Error(`Unable to load graph from "${path}"`);
   return BoardRunner.fromGraphDescriptor(graph);
 };
